+++ apps/webapp/src/features/chat/pages/AiDoctorPage.jsx
@@ -4,82 +4,82 @@ import {
   FaArrowRight,
   FaArrowLeft,
   FaStop,
+  FaPaperPlane,
 } from "react-icons/fa";
 import { useNavigate, useParams } from "react-router-dom";
 import "../../../shared/styles/aiDoctor.css";
 import ChatService from "../services/ChatService";
-import socket from "../../../sockets/socket";
 
 const AiDoctorPage = () => {
-  const [ws, setWs] = useState(null);
   const [isRecording, setIsRecording] = useState(false);
   const [audioChunks, setAudioChunks] = useState([]);
   const [messageInput, setMessageInput] = useState("");
   const [messages, setMessages] = useState([]);
   const [showTyping, setShowTyping] = useState(false);
-  const [messageHistory, setMessageHistory] = useState([]);
-  const navigate = useNavigate();
-
+  const [currentPersona, setCurrentPersona] = useState("general");
+  const [isConnected, setIsConnected] = useState(false);
+  const [patientInfo, setPatientInfo] = useState(null);
   const mediaRecorderRef = useRef(null);
   const chatMessagesRef = useRef(null);
+  const audioRef = useRef(null);
+  const wsRef = useRef(null);
+  const navigate = useNavigate();
   const { patientId } = useParams();
-
   const jwtToken = process.env.REACT_APP_DOCTOR_TOKEN;
   const remoteServerUrl = process.env.REACT_APP_DOCTOR_URL;
-
-  useEffect(() => {
-    connectWebSocket();
-    return () => {
-      if (ws) ws.close();
-    };
-  }, []);
-
-  useEffect(() => {
-    fetchAiDoctorHistory();
-  }, []);
-
-  useEffect(() => {
-    if (chatMessagesRef.current) {
-      chatMessagesRef.current.scrollTop = chatMessagesRef.current.scrollHeight;
-    }
-  }, [messages]);
 
   const fetchAiDoctorHistory = async () => {
     try {
       const response = await ChatService.getAiDoctorChatHistory();
       if (response.status !== 200) {
         navigate("/");
+        return;
       }
-      setMessageHistory(response.data);
-      console.log(response);
+
+      const formattedMessages = response.data.map((msg) => ({
+        id: msg.chat_message_id,
+        text: msg.message_text,
+        sender: msg.sender_id,
+        keywords: msg.keywords || [],
+        time: new Date(msg.timestamp).toLocaleTimeString([], {
+          hour: "2-digit",
+          minute: "2-digit",
+        }),
+      }));
+      setMessages(formattedMessages);
     } catch (err) {
       console.error("Failed to fetch chat history", err);
     }
   };
 
-  const handleRedirect = () => {
-    navigate("/home");
+  const handleFeedback = async (id, type) => {
+    setMessages((prevMessages) =>
+      prevMessages.map((msg) =>
+        msg.id === id ? { ...msg, feedback: type } : msg
+      )
+    );
+
+    await ChatService.sendFeedback({ messageId: id, type });
   };
 
   const connectWebSocket = () => {
-    if (ws) ws.close();
+    if (wsRef.current) {
+      wsRef.current.close();
+    }
 
     try {
       const protocol = remoteServerUrl.startsWith("https") ? "wss:" : "ws:";
       const host = remoteServerUrl.replace(/^https?:\/\//, "");
       const wsUrl = `${protocol}//${host}/chat-final/${patientId}`;
 
+      console.log("Connecting to WebSocket:", wsUrl);
       const websocket = new WebSocket(wsUrl);
-      setWs(websocket);
+      wsRef.current = websocket;
 
       websocket.onopen = () => {
         console.log("WebSocket connection opened");
-        socket.send(
-          JSON.stringify({
-            token: jwtToken,
-            voice: voiceSelect.value
-          })
-        );
+        setIsConnected(true);
+        websocket.send(JSON.stringify({ token: jwtToken }));
       };
 
       websocket.onmessage = (event) => {
@@ -87,13 +87,27 @@ const AiDoctorPage = () => {
           const data = JSON.parse(event.data);
           setShowTyping(false);
 
+          if (data.error) {
+            console.error("Server error:", data.error);
+            addMessage(`Error: ${data.error}`, "system");
+            return;
+          }
+
           if (data.transcription) {
             handleTranscription(data.transcription);
-          } else if (data.response) {
-            handleResponse(data.response);
-          } else if (data.error) {
-            console.error("Server error:", data.error);
-            alert(`Error: ${data.error}`);
+            return;
+          }
+
+          if (data.patient_info) {
+            setPatientInfo(data.patient_info);
+            return;
+          }
+
+          if (data.response) {
+            if (data.current_persona) {
+              setCurrentPersona(data.current_persona);
+            }
+            handleResponse(data.response, data.extracted_keywords, data.audio);
           }
         } catch (error) {
           console.error("Error parsing message:", error);
@@ -104,14 +118,16 @@ const AiDoctorPage = () => {
         console.log(
           `WebSocket closed: Code ${event.code}, Reason: ${event.reason}`
         );
+        setIsConnected(false);
       };
 
       websocket.onerror = (error) => {
         console.error("WebSocket error:", error);
+        setIsConnected(false);
       };
     } catch (error) {
       console.error("Error creating WebSocket:", error);
-      alert(`Error connecting to server: ${error.message}`);
+      addMessage(`Error connecting to server: ${error.message}`, "system");
     }
   };
 
@@ -119,28 +135,52 @@ const AiDoctorPage = () => {
     addMessage(text, "user");
   };
 
-  const handleResponse = async (text) => {
-    const message = text;
-    addMessage(text, "ai");
-    await ChatService.saveAiDoctorResponse(message);
+  const handleResponse = (text, keywords = [], audioData = null) => {
+    addMessage(text, "ai", keywords);
+    if (audioData && audioRef.current) {
+      audioRef.current.src = `data:audio/mp3;base64,${audioData}`;
+      audioRef.current
+        .play()
+        .catch((e) => console.error("Error playing audio:", e));
+    }
   };
 
-  const addMessage = (text, sender) => {
+  const addMessage = async (text, sender, keywords = []) => {
     const newMessage = {
       id: Date.now(),
       text,
       sender,
+      keywords,
       time: new Date().toLocaleTimeString([], {
         hour: "2-digit",
         minute: "2-digit",
       }),
+      feedback: null,
     };
+
     setMessages((prev) => [...prev, newMessage]);
+    try {
+      if (newMessage.sender === "user") {
+        await ChatService.sendAiDoctorMessage(newMessage);
+      }
+      if (newMessage.sender === "ai") {
+        await ChatService.saveAiDoctorResponse(newMessage);
+      }
+    } catch (error) {
+      console.error(
+        `There was an error saving message ${newMessage.id} to the db.`,
+        error
+      );
+    }
   };
 
   const sendMessage = async () => {
     const message = messageInput.trim();
-    if (!message || !ws || ws.readyState !== WebSocket.OPEN) {
+    if (
+      !message ||
+      !wsRef.current ||
+      wsRef.current.readyState !== WebSocket.OPEN
+    ) {
       alert("Not connected to server or message is empty.");
       return;
     }
@@ -149,8 +189,8 @@ const AiDoctorPage = () => {
     setShowTyping(true);
 
     try {
-      ws.send(JSON.stringify({ text: message }));
-      await ChatService.sendAiDoctorMessage(message);
+      console.log(message);
+      wsRef.current.send(JSON.stringify({ text: message }));
     } catch (error) {
       console.error("Error sending message:", error);
       setShowTyping(false);
@@ -185,7 +225,7 @@ const AiDoctorPage = () => {
 
       mediaRecorder.start();
       setIsRecording(true);
-      addMessage("🎤 Recording...", "user");
+      addMessage("Recording...", "system");
     } catch (error) {
       console.error("Microphone error:", error);
       alert(`Microphone error: ${error.message}`);
@@ -202,14 +242,12 @@ const AiDoctorPage = () => {
   const sendAudio = () => {
     const audioBlob = new Blob(audioChunks, { type: "audio/webm" });
     const reader = new FileReader();
-
     setShowTyping(true);
 
     reader.onload = () => {
       const base64Audio = reader.result.split(",")[1];
-
-      if (ws && ws.readyState === WebSocket.OPEN) {
-        ws.send(JSON.stringify({ audio: base64Audio }));
+      if (wsRef.current && wsRef.current.readyState === WebSocket.OPEN) {
+        wsRef.current.send(JSON.stringify({ audio: base64Audio }));
         console.log("Audio sent to server");
       } else {
         console.error("WebSocket not connected");
@@ -220,59 +258,139 @@ const AiDoctorPage = () => {
 
     reader.readAsDataURL(audioBlob);
   };
+
+  const handleBack = () => {
+    navigate("/home");
+  };
+
+  const getPersonaClass = () => `persona-indicator persona-${currentPersona}`;
+  const getPersonaName = () => {
+    const names = {
+      general: "Dr. Ori",
+      psychologist: "Psychologist",
+      dietician: "Dietician",
+    };
+    return names[currentPersona] || "General OPD";
+  };
+
+  useEffect(() => {
+    connectWebSocket();
+
+    return () => {
+      if (wsRef.current) {
+        wsRef.current.close();
+      }
+    };
+  }, []);
+
+  useEffect(() => {
+    if (chatMessagesRef.current) {
+      chatMessagesRef.current.scrollTop = chatMessagesRef.current.scrollHeight;
+    }
+  }, [messages]);
+
+  useEffect(() => {
+    fetchAiDoctorHistory();
+  }, []);
+
   return (
     <div className="chat-container">
+      <audio ref={audioRef} style={{ display: "none" }} />
+
       <div className="chat-header">
-        <button className="back-button">
-          <FaArrowLeft onClick={handleRedirect} />
+        <button className="back-button" onClick={handleBack}>
+          <FaArrowLeft />
         </button>
-        <h2 className="chat-title">Prasha Doctor</h2>
-        <div className="header-placeholder" />
+        <h2 className="chat-title">
+          Healthcare Chat{" "}
+          <span className={getPersonaClass()}>{getPersonaName()}</span>
+        </h2>
+        <div className="connection-status">
+          <span className={isConnected ? "connected" : "disconnected"}>
+            {isConnected ? "Connected" : "Disconnected"}
+          </span>
+        </div>
       </div>
 
-      <div className="chat-messages" ref={chatMessagesRef}>
-        {[...messageHistory, ...messages].length === 0 ? (
-          <div className="empty-chat">
-            <h1>Start a conversation with Prasha Doctor</h1>
+      <div className="chat-body" ref={chatMessagesRef}>
+        {patientInfo && (
+          <div className="sidebar">
+            <div className="patient-info">
+              <h3>Patient Information</h3>
+              <div className="info-content">
+                {patientInfo.name && (
+                  <div className="info-section">
+                    <h4>Basic Info</h4>
+                    <div className="info-item">
+                      <strong>Name:</strong> {patientInfo.name}
+                    </div>
+                    <div className="info-item">
+                      <strong>Gender:</strong> {patientInfo.gender}
+                    </div>
+                  </div>
+                )}
+              </div>
+            </div>
           </div>
-        ) : (
-          [...messageHistory, ...messages].map((msg, index) => (
+        )}
+
+        <div className="messages">
+          {messages.map((msg) => (
             <div
-              key={msg.id || index}
+              key={msg.id}
               className={`message ${
-                msg.sender === "user" ? "user-message" : "ai-message"
+                msg.sender === process.env.REACT_APP_DOCTOR_ID ||
+                msg.sender === "ai"
+                  ? "ai"
+                  : "user"
               }`}
             >
-              <div className="message-text">{msg.message_text || msg.text}</div>
+     
+              <div className="text">{msg.text}</div>
+              <span className="time">{msg.time}</span>
 
-              <div className="message-time">{msg.time}</div>
+              {msg.sender === "ai" && (
+                <div className="feedback-buttons">
+                  <button
+                    onClick={() => handleFeedback(msg.id, "up")}
+                    className={`thumb-button ${
+                      msg.feedback === "up" ? "active" : ""
+                    }`}
+                  >
+                    👍
+                  </button>
+                  <button
+                    onClick={() => handleFeedback(msg.id, "down")}
+                    className={`thumb-button ${
+                      msg.feedback === "down" ? "active" : ""
+                    }`}
+                  >
+                    👎
+                  </button>
+                </div>
+              )}
             </div>
-          ))
-        )}
-
-        {showTyping && (
-          <div className="typing-indicator">
-            <span className="dot" />
-            <span className="dot delay-200" />
-            <span className="dot delay-400" />
-          </div>
-        )}
+          ))}
+          {showTyping && (
+            <div className="message ai">
+              <div className="text">typing...</div>
+            </div>
+          )}
+        </div>
       </div>
 
       <div className="chat-input">
         <input
           type="text"
-          placeholder="Message..."
-          className="input-field"
+          placeholder="Type your message..."
           value={messageInput}
           onChange={(e) => setMessageInput(e.target.value)}
-          onKeyPress={(e) => e.key === "Enter" && sendMessage()}
         />
-        <button className="record-button" onClick={toggleRecording}>
-          {isRecording ? <FaStop className="stop-icon" /> : <FaMicrophone />}
-        </button>
-        <button className="send-button" onClick={sendMessage}>
+        <button onClick={sendMessage}>
           <FaArrowRight />
+        </button>
+        <button onClick={toggleRecording}>
+          {isRecording ? <FaStop /> : <FaMicrophone />}
         </button>
       </div>
     </div>
