+++ apps/webapp/src/shared/styles/aiDoctor.css
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: column;
   height: 100vh;
-  background-color: #fefefe;
+  background-color: #FAF3EC;
   font-family: Arial, sans-serif;
 }
 
@@ -89,25 +89,35 @@
   flex-direction: column;
 }
 
-/* AI messages on left with blue */
 .message.ai {
   align-self: flex-start;
-  background-color: #e0e7ff;
-  color: #1e3a8a;
+  background-color: #E7C39F;
+  color: #000000;
   border-top-left-radius: 4px;
 }
 
-/* User messages on right with green */
 .message.user {
   align-self: flex-end;
-  background-color: #d1fae5;
-  color: #065f46;
+  background-color: #1E0033;
+  color: #FFFFFF;
   border-top-right-radius: 4px;
+}
+
+.avatar {
+  width: 32px;
+  height: 32px;
+  margin-right: 8px;
+}
+
+.avatar img {
+  width: 100%;
+  height: 100%;
+  border-radius: 50%;
 }
 
 .time {
   font-size: 0.7rem;
-  color: #777;
+  color: #707070;
   margin-top: 4px;
   align-self: flex-end;
 }
