+++ apps/webapp/src/shared/styles/homeScreenPage.css
@@ -8,7 +8,7 @@
   padding-bottom: 6rem;
   max-width: 80rem;  
   width: 100%;
-  background-color: white;
+  background-color: #ecf0f1;
   min-height: 100vh;
 }
