+++ apps/webapp/src/shared/styles/otp.css
@@ -60,6 +60,7 @@
   font-size: 24px;
   text-align: center;
   border: 2px solid #ccc;
+  background-color: #D9D9D9;
   border-radius: 8px;
   outline: none;
   transition: border-color 0.3s ease;
