+++ apps/webapp/src/shared/styles/register.css
@@ -1,13 +1,13 @@
 #title {
   padding: 2%;
   border-radius: 10px;
-  background-color: #d9d9d9;
+  background-color: #FFFFFF;
   border: 1px solid #ccc;
 }
 
 #gender {
   padding: 2%;
   border-radius: 10px;
-  background-color: #d9d9d9;
+  background-color: #FFFFFF;
   border: 1px solid #ccc;
 }
